pyproject.toml
[tool.poetry]
name = "api-music-catalog"
version = "0.1.0"
description = ""
authors = ["PavelMoiseev <moiseevpvw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pytest = "^7.2.1"
python-dotenv = "^0.21.1"
drf-yasg = "^1.21.5"
pytest-django = "^4.5.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
